+++ Hazel/src/Platform/Windows/WindowsInput.cpp
@@ -23,17 +23,17 @@ namespace Hazel
         return glfwGetMouseButton(window, static_cast<int>(button)) == GLFW_PRESS;
     }
 
-    double WindowsInput::GetMouseX() const
+    float WindowsInput::GetMouseX() const
     {
         double x, y;
         glfwGetCursorPos(window, &x, &y);
-        return x;
+        return (float)x;
     }
 
-    double WindowsInput::GetMouseY() const
+    float WindowsInput::GetMouseY() const
     {
         double x, y;
         glfwGetCursorPos(window, &x, &y);
-        return y;
+        return (float)y;
     }
 }
